fix: deploy script will only do copy command if build is success

+++ scripts/post-deploy-client.sh
@@ -6,5 +6,7 @@ echo "[3/5] installing dependencies..."
 yarn --prefer-offline --silent
 echo "[4/5] building..."
 touch .cache && rm -r ./.cache && yarn run build:client
-echo "[5/5] copying..."
-rm -r /var/www/leavemealone/source/public && mv /var/www/leavemealone/next/source/public/ /var/www/leavemealone/source
+if [ $? -eq 0 ]; then
+    echo "[5/5] copying..."
+    rm -r /var/www/leavemealone/source/public && mv /var/www/leavemealone/next/source/public/ /var/www/leavemealone/source
+fi
